[spec] Removing uncessary tests

+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@
 ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
+require 'rubygems'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
@@ -26,7 +27,7 @@ RSpec.configure do |config|
   # config.use_transactional_fixtures = true
   config.before :each do
     ActiveRecord::Base.send(:descendants).each do |klass|
-      klass.destroy_all
+      klass.destroy_all if klass
     end
   end
 end
